Update TweetsList to separate TweetCard component
Moved TweetCard to its own file and updated the test
